+++ script/tool/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fixes changelog validation when reverting to a `NEXT` state.
 - Fixes multiplication of `--force` flag when publishing multiple packages.
+- Checks for template boilerplate in `readme-check`.
 
 ## 0.8.5